Changed float parameters to int16 in 'cli.c'.
Fixed buffer length used for 'ftoa'.

Got rid of magic numbers, replaced 'digitalIdlePercent' with 'digitalIdleValue'.

+++ src/main/telemetry/telemetry.h
@@ -38,8 +38,8 @@ typedef enum {
 } frskyUnit_e;
 
 typedef struct telemetryConfig_s {
-    float gpsNoFixLatitude;
-    float gpsNoFixLongitude;
+    int16_t gpsNoFixLatitude;
+    int16_t gpsNoFixLongitude;
     uint8_t telemetry_switch;               // Use aux channel to change serial output & baudrate( MSP / Telemetry ). It disables automatic switching to Telemetry when armed.
     uint8_t telemetry_inversion;            // also shared with smartport inversion
     uint8_t sportHalfDuplex;
